Inject Transformed Scripts into Original HTML Test (#4)
* Added transformation and tests for basic synchronous test functions.

* Transformation function and tests for basic synchronous functions.

Each transformation is a Babel plugin, organized by the change to
parameters needed to transform the function.

* Removed merge artifacts

* Addressed PR comments

* Inject scripts into HTML using posthtml

* Added script nodes with correct src

* Async/await in injectScripts

* Refactored to eliminate files, injectScripts into HTML

This is a large commit, but mostly refactoring. Instead of writing to
the file system and having to remember all of the paths and naming
conventions just to read the file in the next step, this code now returns strings and
objects where needed, allowing index.js to route everything properly. This makes
testing each individual component easier as well.
Commit also includes injectScripts.js and relevant tests to inject
transformed .js scripts back into the correct position in HTML.

* PR comments: removed tempDir

* PR comments: removed tempDir

* Indentation change